+++ src/components/error-fallback/index.ts
@@ -1,0 +1,1 @@
+export * from './ErrorFallback';
