+++ src/components/index.ts
@@ -1,6 +1,7 @@
 export * from './button';
 export * from './call-to-action';
 export * from './dropdown';
+export * from './error-fallback';
 export * from './form';
 export * from './header';
 export * from './modal';
