--- src/components/not-found-buttons/index.ts
@@ -1,1 +1,0 @@
-export * from './NotFoundButtons';
